Made tags only accessible per user

+++ src/tags/tags.controller.ts
@@ -16,16 +16,21 @@ import {
 import { ZodValidationPipe } from "src/validation.pipe";
 import { CreateTagDto, createTagSchema } from "./dto/create-tag.dto";
 import { TagsService } from "./tags.service";
+import { AuthIdGuard } from "src/auth.guard";
+import { ApiBearerAuth } from "@nestjs/swagger";
 
 @Controller("tags")
 export class TagsController {
   constructor(private readonly tagsService: TagsService) {}
 
+  @ApiBearerAuth()
   @Post()
+  @UseGuards(AuthIdGuard)
   create(
     @Body(new ZodValidationPipe(createTagSchema)) createTagDto: CreateTagDto,
+    @Req() request: Request,
   ) {
-    return this.tagsService.create(createTagDto).catch((err) => {
+    return this.tagsService.create(createTagDto, request["user"]?.id).catch((err) => {
       if (err.name === "SequelizeUniqueConstraintError") {
         throw new ConflictException("Tag already exists");
       } else {
@@ -34,13 +39,31 @@ export class TagsController {
     });
   }
 
+  @ApiBearerAuth()
   @Get()
-  findAll() {
-    return this.tagsService.findAll();
+  @UseGuards(AuthIdGuard)
+  findAll(
+    @Req() request: Request,
+  ) {
+    return this.tagsService.findAll(request["user"]?.id);
   }
 
+  @ApiBearerAuth()
   @Delete(":id")
-  async remove(@Param("id", ParseIntPipe) id: number) {
+  @UseGuards(AuthIdGuard)
+  async remove(
+    @Param("id", ParseIntPipe) id: number,
+    @Req() request: Request,
+  ) {
+    const tag = await this.tagsService.findOne(id);
+
+    if (!tag) {
+      throw new NotFoundException("Tag with this id does not exist");
+    }
+
+    if (tag.userId !== request["user"]?.id) {
+      throw new ForbiddenException("You cannot delete tags from other users");
+    }
 
     const ret = await this.tagsService.remove(id).catch((err) => {
       if (err.name === "SequelizeForeignKeyConstraintError") {
